+++ src/SFML/Window/iOS/VideoModeImpl.mm
@@ -50,8 +50,8 @@ std::vector<VideoMode> VideoModeImpl::getFullscreenModes()
 VideoMode VideoModeImpl::getDesktopMode()
 {
     CGRect bounds = [[UIScreen mainScreen] bounds];
-    float backingScale = [SFAppDelegate getInstance].backingScaleFactor;
-    return VideoMode(bounds.size.width * backingScale, bounds.size.height * backingScale);
+    double backingScale = [SFAppDelegate getInstance].backingScaleFactor;
+    return VideoMode(static_cast<unsigned int>(bounds.size.width * backingScale), static_cast<unsigned int>(bounds.size.height * backingScale));
 }
 
 } // namespace priv
